General edits
General suggestions for edits. In the terminology, "Participating Countries" are defined as both Issuers and Verifiers. But in the rest of the text, there was Issuing Countries and Verifying Countries. In the EU context, they are the same. So I suggest just using the term "Participating Countries" to imply both.